Create W02_P3.java
a program to demonstrate the use of this keyword. Create a class Rectangle with attributes length and breadth. Write a constructor to initialize these values. Also, define a method area() to return the area of the rectangle.

Input:
Two integers: length and breadth of the rectangle.

Output:
Print the area of the rectangle