Create W02_P5.java
Java program to demonstrate the concept of encapsulation by modeling a Book object. The program should include methods for setting and getting the title and author of the book, and a method to display the details of the book.

Task:

Create a class Book with the following private attributes:

title (String)
author (String)
Include the following methods in the Book class:

setTitle(String title) to set the title of the book.
setAuthor(String author) to set the author of the book.
getTitle() to return the title of the book.
getAuthor() to return the author of the book.
displayDetails() to print the details of the book in the format
Title: <title>  
Author: <author>  
In the main() method, create an object of the Book class, set the title and author, and then call the displayDetails() method to print the book details.